fix: Centralize path validation in load_resources() and add type hint
- Move _assert_safe_path() from app.py into rag.py so both CLI and UI
  callers benefit from the same safety checks (path traversal, symlink,
  world-writable) without duplication
- Call _assert_safe_path() inside load_resources() before deserialization
- Add docs: list[Document] type hint to format_docs()
